Comment out model references. They need JPA notations for work

+++ e-innsynv2/src/main/java/com/dc19einnsyn/einnsynv2/Models/Dokumentobjekt.java
@@ -14,7 +14,7 @@ public class Dokumentobjekt {
 
     @Id
     @GeneratedValue
-    private int dokumentObjektId;
+    private long dokumentObjektId;
 
     private String versjonsnummer;
     private String variantformat;
