+++ e-innsynv2/src/main/java/com/dc19einnsyn/einnsynv2/Models/Korrespondansepart.java
@@ -14,7 +14,7 @@ public class Korrespondansepart {
 
     @Id
     @GeneratedValue
-    private int korrespondansePartId;
+    private long korrespondansePartId;
 
     private String korrespondanseparttype;
     private String korrespondansepartNavn;
